slight refactoring of ./p2p with improved semantics
